all 7 commands code written successfully

+++ src/Accounts.java
@@ -1,15 +1,15 @@
 public class Accounts {
-    private int money;
+    private double money;
 
-    public Accounts(int money) {
+    public Accounts(double money) {
         this.money = money;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
 }
